backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
apscheduler==3.10.4
jinja2==3.1.2
python-dateutil==2.8.2
google-cloud-firestore==2.15.0
google-auth==2.27.0
firebase-admin==6.5.0
httpx==0.27.0
phonenumbers==9.0.24
pyjwt==2.11.0
mock-firestore==0.11.0
slowapi==0.1.9
cryptography==42.0.5

pytest==7.4.0
